+++ examples/Example.Api/tests/Example.Api.Tests.Component.LightBDD.xUnit2/Infrastructure/ConfiguredLightBddScopeAttribute.cs
@@ -32,7 +32,7 @@ internal class ConfiguredLightBddScopeAttribute : LightBddScopeAttribute
                 SeparateSetup = true,
             };
 
-        configuration.ReportWritersConfiguration().CreateStandardReportsWithDiagrams(reportOptions);
+        configuration.CreateStandardReportsWithDiagrams(reportOptions);
 
         // To stop the output repeating the step name for each step
         configuration.ProgressNotifierConfiguration().Clear().Append(new ConfigurableProgressNotifier());